fastapi-api/requirements.txt
fastapi==0.62.0
uvicorn==0.12.3
motor==2.3.0
pytest==6.1.2
python-dotenv==0.15.0
requests==2.25.0